EA Foundation complete, for now...
Completed the EA foundation for the ADDI instruction. There are still a
couple of modes that need to be implemented (Address register direct and
immediate data) but we will get to those as we get to them.

I also need to look into handling EA decoding for a SOURCE, but this
shouldn't be to much extra work. Just some wrappers to output slightly
different from DESTINATION.

Should also be able to read in two consecutive ADDI instruction, try
with the input file.