Update with start and end address
Successfully implemented the request for user to enter start and end
addresses. Also finished the implementation of ASCII to HEX conversion.

Only thing left to do is finish the HEX to ASCII conversion (which is
exactly like the ASCII to HEX, but I'd need to know which registers I'm
working with), and also printing out the memory addresses of the
op-codes. And then we should be done with I/O!!!!