Added manual commas, more EA_BCHG_D work.
Removed the coupling of commas with registers, it is now every EA funcs
responsibility to manually insert the comma where appropriate.

Added the functionality to output a register without having always
having to go to next instruction. To output a register and RTS back,
flag D6 with a #$01.

Now specify the size of the absolute address mode (01 = word, 10 = long)
into register D1, so that the hextoascii converter works properly.

More general work in EA_BCHG_D, nearly done...