Implemented the continue question
If user says they don't want to continue, they will be asked if they
want to do the disassembler again. If they don't, then it quits. If they
wanted to continue, it'll proceed as usual; it does this by resetting
the counter. Once the end address is reach though, it'll automatically
proceed back to the intro again, asking if the user wants to do another
test run.