BCHG(dyn) = complete. Changed NewLine behavior
The BCHG(dynamic) opcode should be be decoding properly, for the most
part. Had to handle a couple tricky cases for it.

I changed the newline behavior of the program. Now every time we enter
the main loop 'grab_next_op', the first thing we do is print a new line.

This removes the coupling between newlines for registers and allows all
of our newlines to take place in one part of the program...